% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{calculateNormalizationBySpikeinGLM}
\alias{calculateNormalizationBySpikeinGLM}
\title{Fit GLM to spike-in read counts to extract sequencing depth and
cross-contamination rate for all samples}
\usage{
calculateNormalizationBySpikeinGLM(spikeinCounts)
}
\arguments{
\item{spikeinCounts}{rCubeExperiment Object containing read counts for spike-ins}
}
\value{
rCubeExperiment Object for spike-ins with updated information
}
\description{
Fit GLM to spike-in read counts to extract sequencing depth and
cross-contamination rate for all samples
}
\author{
Carina Demel
}
