% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countJunctions.R, R/countSpikeins.R
\name{countJunctions}
\alias{countJunctions}
\alias{countSpikeins}
\alias{countFeatures}
\title{Counting reads for features}
\usage{
countJunctions(experimentalSetup, scanBamParam = Rsamtools::ScanBamParam(flag
  = Rsamtools::scanBamFlag(isSecondaryAlignment = FALSE)), BPPARAM = NULL,
  ncores = 1, verbose = FALSE)

countSpikeins(experimentalSetup, scanBamParam = Rsamtools::ScanBamParam(flag =
  Rsamtools::scanBamFlag(isSecondaryAlignment = FALSE)), BPPARAM = NULL,
  verbose = FALSE)

countFeatures(experimentalSetup, scanBamParam = Rsamtools::ScanBamParam(flag =
  Rsamtools::scanBamFlag(isSecondaryAlignment = FALSE)), BPPARAM = NULL,
  verbose = FALSE)
}
\arguments{
\item{experimentalSetup}{An empty \code{rCubeExperiment} object for 
spikeins/junctions, see \code{\link{setupExperimentSpikeins}} and 
\code{\link{setupExperiment}}.}

\item{scanBamParam}{Parameter to specify which fields are imported from a BAM
file.}

\item{BPPARAM}{An instance of a \code{BiocParallelParam} class, e.g., 
\code{\link{MulticoreParam}}, \code{\link{SnowParam}}, 
\code{\link{DoparParam}}.}

\item{ncores}{Number of cores to be used for parallel computation.}

\item{verbose}{If true, status messages are printed.}
}
\value{
An updated \code{rCubeExperiment} container, with spike-in read 
counts.
}
\description{
\code{countSpikeins} provides a framework to count reads for 
artificial spikeins, \code{countJunctions} to count reads for junctions.
}
\examples{
data(junctions)
bamfiles <- list.files(system.file("extdata/TimeSeriesExample/", package='rCube'), 
pattern=".bam$", full.names=TRUE)
# first setup experiment
junctionCounts <- setupExperiment(junctions, files=bamfiles)
# then count reads on junctions
junctionCounts <- countJunctions(junctionCounts, 
BPPARAM=BiocParallel::MulticoreParam(1))
data("spikeins")
data("spikeinLabeling")
folder <- system.file("extdata/Jurkat", package="rCube")
bamfiles <- list.files(folder, pattern="*.bam$", full.names=TRUE)
spikeinCounts <- setupExperimentSpikeins(rows=spikeins, 
files=bamfiles, labelingState=spikeinLabeling)
assay(spikeinCounts)
spikeinCounts <- countSpikeins(spikeinCounts)
assay(spikeinCounts)
data(exampleExons)
folder <- system.file("extdata/Jurkat", package="rCube")
bamfiles <- list.files(folder, pattern="*.bam$", full.names=TRUE)
exonCounts <- setupExperiment(exampleExons, designMatrix=NULL, files=bamfiles)
assay(exonCounts)
exonCounts <- countFeatures(exonCounts)
assay(exonCounts)
}
\seealso{
\code{\link[Rsamtools]{ScanBamParam}}

\code{\link{BiocParallelParam}}
}
\author{
Leonhard Wachutka
}
