% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countJunctions.R
\name{countJunctions}
\alias{countJunctions}
\title{count Donor, Acceptor and Junction reads}
\usage{
countJunctions(experimentalSetup, filter = NULL, BPPARAM = NULL,
  ncores = 1, verbose = FALSE)
}
\arguments{
\item{rows}{GRanges}

\item{designMatrix}{dgM}
}
\value{
An empty rCubeExperiment container
}
\description{
count Donor, Acceptor and Junction reads
}
\examples{

}
\author{
Leonhard Wachutka
}
