% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCube-package.R
\docType{data}
\name{countMatrixJunctions}
\alias{countMatrixJunctions}
\title{Count matrix for Junctions retrieved from TimeSeriesExample BAM files}
\description{
Count matrix for Junctions retrieved from TimeSeriesExample BAM files
}
\author{
Leonhard Wachutka, Carina Demel, Julien Gagneur
}
\keyword{data}
