% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCube-package.R
\docType{data}
\name{countMatrixSpikeins}
\alias{countMatrixSpikeins}
\title{Count matrix for spike-ins retrieved from TimeSeriesExample BAM files}
\description{
Count matrix for spike-ins retrieved from TimeSeriesExample BAM files
}
\author{
Leonhard Wachutka, Carina Demel, Julien Gagneur
}
\keyword{data}
