% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countSpikeins.R
\name{countSpikeins}
\alias{countSpikeins}
\title{count Spikeins}
\usage{
countSpikeins(experimentalSetup, scanBamParam = ScanBamParam(flag =
  scanBamFlag(isSecondaryAlignment = FALSE)), BPPARAM = NULL,
  verbose = FALSE)
}
\arguments{
\item{rows}{GRanges}

\item{designMatrix}{dgM}
}
\value{
An empty rCubeExperiment container
}
\description{
count Spikeins
}
\examples{

}
\author{
Leonhard Wachutka
}
