% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createConstitutiveFeaturesGRangesFromGRanges.R
\name{createConstitutiveFeaturesGRangesFromGRanges}
\alias{createConstitutiveFeaturesGRangesFromGRanges}
\title{A function to create constitutive features (i.e. exons or introns) from an
annotation file. All exonic/intronic bases are considered constitutive, when
they belong to all transcript isoform of the corresponding gene.}
\usage{
createConstitutiveFeaturesGRangesFromGRanges(granges, ncores = 2)
}
\arguments{
\item{granges}{A GRanges annotation file with exons/introns. Needs columns 
"gene_id" and "transcript_id"}

\item{ncores}{Number of cores available for parallel computation}
}
\value{
Returns a GRanges object with constitutive exons.
}
\description{
A function to create constitutive features (i.e. exons or introns) from an
annotation file. All exonic/intronic bases are considered constitutive, when
they belong to all transcript isoform of the corresponding gene.
}
\examples{
# Gencode annotation of MYC gene
data(example.exons)
constitutive.exons = createConstitutiveFeaturesGRangesFromGRanges(example.exons, 1)
}
\author{
Carina Demel
}
