% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createJunctionGRangesFromBam.R
\name{createJunctionGRangesFromBam}
\alias{createJunctionGRangesFromBam}
\title{Title}
\usage{
createJunctionGRangesFromBam(bamFiles, support = 10, ncores = 2,
  BPPARAM = NULL)
}
\arguments{
\item{bamFiles}{A string Vector containing all bam files}

\item{support}{min number of supporting reads}

\item{ncores}{Number of cores available for parallel computation}

\item{BPPARAM}{see BiocParallel}
}
\value{
Returns a GRanges object with junctions.
}
\description{
Title
}
\examples{
# Gencode annotation of MYC gene
data(example.exons)
constitutive.exons = createConstitutiveFeaturesGRangesFromGRanges(example.exons, 1)
}
\author{
Leonhard Wachutka
}
