% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCube-package.R
\docType{data}
\name{designMatrix}
\alias{designMatrix}
\title{The desginMatrix for the Quick-Start example in the vignette}
\description{
The desginMatrix for the Quick-Start example in the vignette
}
\author{
Leonhard Wachutka, Carina Demel, Julien Gagneur
}
\keyword{data}
