% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateRateByFristOrderKineticsSingle.R
\name{estimateRateByFirstOrderKineticsSingle}
\alias{estimateRateByFirstOrderKineticsSingle}
\title{Estimation of synthesis and degradation rates (half-lives) based on
4sU-labeled and total RNA-seq read counts.}
\usage{
estimateRateByFirstOrderKineticsSingle(featureCounts, spikeinCounts,
  replicate = NULL, BPPARAM = NULL)
}
\arguments{
\item{featureCounts}{A \code{rCubeExperiment} object with feature count table}

\item{spikeinCounts}{A \code{rCubeExperiment} object with spike-in count table and
sequencing.depth and crossContaminationamination information (e.g. output from 
\code{\link[estimateSizeDispersions]{estimateSizeDispersions}})}

\item{replicate}{list of character/factor vectors for which replicate or 
combination of replicates the results should be computed. If \code{NULL},
estimates for each replicate individually and all combinations of replicates
will be calculated.}

\item{BPPARAM}{An instance of a \code{BiocParallelParam} class, e.g., 
\code{\link{MulticoreParam}}, \code{\link{SnowParam}}, \code{\link{DoparParam}}.}
}
\value{
Returns a \code{rCubeRates} object with estimated synthesis and degradation
rates for each feature and sample and the specified replicate combinations
}
\description{
Estimation of synthesis and degradation rates (half-lives) based on
4sU-labeled and total RNA-seq read counts.
}
\examples{
## estimate sequencing depths and cross-contamination values from spike-ins
data(spikeinCounts)
data(geneCounts)
## spikeinCounts <- estimateSizeFactors(geneCounts, spikeinCounts, method = "spikeinGLM") ##TODO return type
spikeinCounts <- calculateNormalizationBySpikeinGLM(spikeinCounts)
## estimate Dispersions for all genes
geneCounts <- estimateSizeDispersions(geneCounts, method='DESeqDispMAP')
## estimate synthesis and degradation rates for individual replicates and combination
rates <- estimateRateByFirstOrderKineticsSingle(geneCounts, spikeinCounts)

}
\seealso{
\code{\link{BiocParallelParam}}
}
\author{
Carina Demel
}
