% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateRateByFristOrderKinetics.R
\name{estimateRateByFristOrderKinetics}
\alias{estimateRateByFristOrderKinetics}
\title{Title}
\usage{
estimateRateByFristOrderKinetics(featureCounts, replicate,
  method = c("series", "single"), BPPARAM = NULL)
}
\value{
A \code{rCubeRates} object with rates
}
\description{
Title
}
\author{
Leonhard Wachutka
}
