% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateSizeDispersions.R
\name{estimateSizeDispersions}
\alias{estimateSizeDispersions}
\title{Title}
\usage{
estimateSizeDispersions(experiment, method = c("DESeqDispMAP", "DESeqDispFit",
  "DESeqDispGeneEst", "Replicate"))
}
\value{
rCubeExperiment Object with updated metadata information
}
\description{
Title
}
\author{
Leonhard Wachutka
}
