% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{estimateSizeFactors}
\alias{estimateSizeFactors}
\title{Estimation of sample-specific normalization factors}
\usage{
estimateSizeFactors(featureCounts, spikeinCounts, method = c("spikeinGLM",
  "spikeinMean", "spikeinMedian"))
}
\arguments{
\item{featureCounts}{A \code{rCubeExperiment} object containing read
counts for features (genes, exons, introns, junctions, ...), see constructor 
function \code{\link{setupExperiment}}.}

\item{spikeinCounts}{A \code{rCubeExperiment} object containing read 
counts, length and labeling status for spike-ins, as well as sample 
information, see constructor function \code{\link{setupExperimentSpikeins}}.}

\item{method}{string specifying which normalization method should be used,
one of \code{('spikeinGLM', 'spikeinMean', 'jointModel')}.}
}
\value{
An updated \code{rCubeExperiment} object for features with sequencing
 depth values included in \code{colData} and updated metadata information.
}
\description{
\code{estimateSizeFactors} provides a wrapper function to
estimate sample-specific size-factors by three different types. The 
'spikeinGLM' method fits a GLM to spike-in read counts to extract sequencing 
depth values and cross-contamination rates for all samples (default).

The 'spikeinMean' and 'spikeinMedian' methods use the mean or median counts 
across labeled spike-ins per sample to determine sample-specific size factors.
For the estimation of cross-contamination in labeled samples, the ratio of 
counts from unlabeled spike-ins to counts from labeled spike-ins is used as
an approximation.

Additional informations for the fit can be found in the metadata of the 
resulting \code{rCubeExperiment}.
}
\examples{
# fitting a GLM to spike-in read counts
data(exonCounts)
data(spikeinCounts)
exonCounts <- estimateSizeFactors(exonCounts, spikeinCounts, method="spikeinGLM")
colData(exonCounts)
}
\author{
Carina Demel, Leonhard Wachutka
}
