% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCube-package.R
\docType{data}
\name{exampleExons}
\alias{exampleExons}
\title{GENCODe annotated exons for MYC}
\description{
GENCODe annotated exons for MYC
}
\author{
Leonhard Wachutka, Carina Demel, Julien Gagneur
}
\keyword{data}
