% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCube-package.R
\docType{data}
\name{exonRates}
\alias{exonRates}
\title{\code{rCubeRates} container for example exons}
\description{
\code{rCubeRates} container for example exons
}
\author{
Leonhard Wachutka, Carina Demel, Julien Gagneur
}
\keyword{data}
