% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCube-package.R
\docType{data}
\name{geneCounts}
\alias{geneCounts}
\title{\code{rCubeExperiment} container for example genes including read counts}
\description{
\code{rCubeExperiment} container for example genes including read counts
}
\author{
Leonhard Wachutka, Carina Demel, Julien Gagneur
}
\keyword{data}
