% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCube-package.R
\docType{data}
\name{geneRates}
\alias{geneRates}
\title{\code{rCubeRates} container for example genes}
\description{
\code{rCubeRates} container for example genes
}
\author{
Leonhard Wachutka, Carina Demel, Julien Gagneur
}
\keyword{data}
