% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCube-package.R
\docType{data}
\name{junctions}
\alias{junctions}
\title{Junction annotation retrieved from TimeSeriesExample BAM files}
\description{
Junction annotation retrieved from TimeSeriesExample BAM files
}
\author{
Leonhard Wachutka, Carina Demel, Julien Gagneur
}
\keyword{data}
