% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaics.R
\docType{data}
\name{mosaics}
\alias{mosaics}
\title{Mosaics list}
\format{A list containing rubicubism mosaics. Projects are 'scanned' from https://www.youcandothecube.com/build-mosaics-with-rubiks-cubes/mosaic-templates}
\usage{
data(mosaics)
}
\description{
A list containing rubicubism mosaics.
}
\details{
\itemize{
  \item annefrank
  \item einstein
  \item monalisa
  \item nelson
  \item sittingbull
  \item tesla
  \item vangogh
}
}
\keyword{datasets}
