% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cube.R
\name{plot.cube}
\alias{plot.cube}
\title{Plotting cube}
\usage{
\method{plot}{cube}(x, ...)
}
\arguments{
\item{x}{- cube object}

\item{...}{- not used}
}
\value{
plot
}
\description{
Plots cube in 2D
}
\examples{
cube <- createCube()
plot(cube)

}
