% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3dFlat.R
\name{plot3dFlat}
\alias{plot3dFlat}
\title{Plotting cube in 2D}
\usage{
plot3dFlat(cube)
}
\arguments{
\item{cube}{- cube object}
}
\value{
plot
}
\description{
Plotting cube in 2D, but holds 3D advantages
}
\examples{
cube <- createCube()
plot3dFlat(cube) # generates plot of solved cube
# Plotting 'checkerboard' pattern using pipe:
require(magrittr)
createCube() \%>\% twistCube("(LLFFRRBB) x3") \%>\% plot3dFlat()
}
