% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plotSpikeinCountsVsSample}
\alias{plotSpikeinCountsVsSample}
\title{Diagnostic plots}
\usage{
plotSpikeinCountsVsSample(spikeinCounts)
}
\arguments{
\item{spikeinCounts}{A \code{rCubeExperiment} object containing read 
counts and labeling status for spike-ins, as well as sample information,
see constructor function \code{\link{setupExperimentSpikeins}}.}
}
\value{
A \code{trellis} object.
}
\description{
\code{plotSpikeinCountsVsSample} visualizes spike-in read counts
per sample to check if experiment worked and labeled spike-ins are enriched,
but unlabeled spike-ins are depleted in labeled samples, whereas all spike-ins
should be present to a similar extend in total RNA-seq samples.
}
\examples{
data(spikeinCounts)
plotSpikeinCountsVsSample(spikeinCounts)
}
\author{
Carina Demel
}
