% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positions.R
\docType{data}
\name{positions}
\alias{positions}
\title{Example dataset}
\format{A data frame with popular positions of cubes}
\usage{
data(positions)
}
\description{
A dataset containing famous cubes positions. The variables are as follows:
}
\details{
\itemize{
  \item n: size of cube
  \item name: pattern name
  \item moves: moves to make
}
}
\keyword{datasets}
