% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cube.R
\name{print.cube}
\alias{print.cube}
\title{Printing cube}
\usage{
\method{print}{cube}(x, ...)
}
\arguments{
\item{x}{- cube object}

\item{...}{- not used}
}
\value{
plain text
}
\description{
Prints cube in console
}
\examples{
cube <- createCube()
print(cube)

}
