% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCubeExperiment.R, R/setupExperiment.R
\docType{class}
\name{rCubeExperiment-class}
\alias{rCubeExperiment-class}
\alias{setupExperimentSpikeins}
\alias{setupExperiment}
\title{rCubeExperiment object and constructors}
\usage{
setupExperimentSpikeins(rows, designMatrix = NULL, files = NULL,
  length = NULL, labelingState, counts = NULL)

setupExperiment(rows, designMatrix = NULL, files = NULL)
}
\arguments{
\item{rows}{GRanges containing spike-in/gene/exon/... annotation}

\item{designMatrix}{an optional data.frame containing sample information.
Either \code{designMatrix} or \code{files} needs to be present.}

\item{files}{An optional vector of sample file names in the format 
'condition_L|T_labelingTime_Replicate.bam'. Either \code{designMatrix}
or \code{files} needs to be present.}

\item{length}{numeric vector with the length of each spike-ins}

\item{labelingState}{factor vector containing L and U indicating if a spikein
was labeled or unlabeled}

\item{counts}{an optional matrix with spike-in read counts}
}
\value{
An (empty) \code{rCubeExperiment} container
}
\description{
\code{rCubeExperiment} is a subclass of 
\code{RangedSummarizedExperiment}, used to store read counts, annotations, and
sample metadata.

\code{setupExperimentSpikeins} generates a \code{rCubeExperiment}
object for spike-ins, which can be used for counting and which is necessary
for normalization and estimation steps.

\code{setupExperiment} generates an empty \code{rCubeExperiment}
container.
}
\examples{
data(spikeins)
data(spikeinLabeling)
spikeinLengths <- width(spikeins)
data(designMatrix)
spikeinCounts <- setupExperimentSpikeins(rows=spikeins, 
designMatrix=designMatrix, length=spikeinLengths, 
labelingState=spikeinLabeling)
data(junctions)
bamfiles <- list.files(system.file("extdata/K562", package='rCube'), pattern=".bam$", full.names=TRUE)
junctionCounts <- setupExperiment(junctions, files=bamfiles)
}
\seealso{
\code{\link[SummarizedExperiment]{SummarizedExperiment}}
}
\author{
Carina Demel, Leonhard Wachutka

Leonhard Wachutka
}
