% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateRateByFirstOrderKinetics.R,
%   R/rCubeRates.R
\docType{class}
\name{createRResultCubeRates}
\alias{createRResultCubeRates}
\alias{createRResultCubeRatesExtended}
\alias{rCubeRates-class}
\title{rCubeRate object and constructors}
\usage{
createRResultCubeRates(featureCounts, replicate)

createRResultCubeRatesExtended(featureCounts, replicate)
}
\arguments{
\item{featureCounts}{a \code{rCubeExperiment} object with set 'condition'
column in \code{colData}.}

\item{replicate}{a vector of integers and strings indicating single replicates/
combination of replicates for which the results should be calculated.}
}
\value{
Returns an empty rCubeRate object.
}
\description{
\code{createRResultCubeRates} is a constructor for 
\code{rCubeExperiment} objects which contains slots for synthesis and 
degradation rates for all replicates/combination of replicates.

\code{createRResultCubeRatesExtended} is a constructor for 
\code{rCubeExperiment} objects which contains slots for synthesis and 
degradation rates, as well as half.lives, labeled and unlabeled RNA amount 
estimates for all replicates/combination of replicates.

\code{rCubeRate} is a subclass of 
\code{RangedSummarizedExperiment}, used to store estimation results such as
synthesis and degradation rates
}
\examples{
data(geneCounts)
rates <- createRResultCubeRates(geneCounts, replicate=c(1, 2, '1:2'))
rates
#' data(geneCounts)
rates <- createRResultCubeRatesExtended(geneCounts, replicate=c(1, 2, '1:2'))
rates
}
\author{
Leonhard Wachutka
}
