% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scramble.R
\name{scramble}
\alias{scramble}
\title{Scrambling cube}
\usage{
scramble(cube, times = 0)
}
\arguments{
\item{cube}{- cube object to scramble}

\item{times}{- how many random moves should be done on cube. Default is 0 which means N*10 moves where N is the size of the cube.}
}
\value{
cube
}
\description{
Scrambling cube
}
\examples{
cube <- createCube()
set.seed(1)
cube <- scramble(cube)
bigcube <- createCube(N = 15)
set.seed(1)
bigcube <- scramble(bigcube)
}
