% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupExperiment.R
\name{setupExperiment}
\alias{setupExperiment}
\title{Generate empty rCubeExperiment Object}
\usage{
setupExperiment(rows, designMatrix)
}
\arguments{
\item{rows}{GRanges}

\item{designMatrix}{dgM}
}
\value{
An empty rCubeExperiment container
}
\description{
Generate empty rCubeExperiment Object
}
\examples{

}
\author{
Leonhard Wachutka
}
