% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupExperiment.R
\name{setupExperimentSpikeins}
\alias{setupExperimentSpikeins}
\title{Generate empty rCubeExperiment Object for spike-ins, which can be used for counting}
\usage{
setupExperimentSpikeins(rows, designMatrix = NULL, files = NULL,
  length = NULL, labelingState)
}
\arguments{
\item{rows}{GRanges containing spike-in annotation}

\item{length}{numeric vector with the length of each spike-ins}

\item{labelingState}{factor vector or logical###??}

\item{path}{string}
}
\value{
An empty rCubeExperiment container
}
\description{
Generate empty rCubeExperiment Object for spike-ins, which can be used for counting
}
\examples{
data(spikein.labeling)
data(spikein.lengths)
spikein.counts <- setupExperimentSpikeins(rows = spikeins, path = "", 
length = spikein.lengths, labelingState = spikein.labeling)
}
\author{
Carina Demel, Leonhard Wachutka
}
