% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCube-package.R
\docType{data}
\name{spikeinCounts}
\alias{spikeinCounts}
\title{\code{rCubeExperiment} container for spike-ins including example read counts}
\description{
\code{rCubeExperiment} container for spike-ins including example read counts
}
\author{
Leonhard Wachutka, Carina Demel, Julien Gagneur
}
\keyword{data}
