% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCube-package.R
\docType{data}
\name{spikeinGenome}
\alias{spikeinGenome}
\title{Biostrings object for artifical spike-ins containing fasta sequences}
\description{
Biostrings object for artifical spike-ins containing fasta sequences
}
\author{
Leonhard Wachutka, Carina Demel, Julien Gagneur
}
\keyword{data}
