% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCube-package.R
\docType{data}
\name{spikeinLabeling}
\alias{spikeinLabeling}
\title{Labeling status of artificial spike-ins sequences}
\description{
Labeling status of artificial spike-ins sequences
}
\author{
Leonhard Wachutka, Carina Demel, Julien Gagneur
}
\keyword{data}
