% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCube-package.R
\docType{data}
\name{spikeinLengths}
\alias{spikeinLengths}
\title{Length of artificial spike-ins sequences}
\description{
Length of artificial spike-ins sequences
}
\author{
Leonhard Wachutka, Carina Demel, Julien Gagneur
}
\keyword{data}
