% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCube-package.R
\docType{data}
\name{spikeins}
\alias{spikeins}
\title{Genomic ranges object for artifical spike-ins, which can be used for counting}
\description{
Genomic ranges object for artifical spike-ins, which can be used for counting
}
\author{
Leonhard Wachutka, Carina Demel, Julien Gagneur
}
\keyword{data}
