% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeRates.R
\name{summarizeRates}
\alias{summarizeRates}
\title{Title}
\usage{
summarizeRates(featureRates, topLevelFeatures, by, featureCounts = NULL)
}
\arguments{
\item{featureRates}{An \code{rCubeRate object}, see constructor function 
\code{\link{createRResultCubeRates}}.}

\item{topLevelFeatures}{\code{GRanges} annotation}

\item{by}{grouping}

\item{featureCounts}{A \code{rCubeExperiment} object, see constructor function 
\code{\link{setupExperiment}}.}
}
\value{
Returns a \code{rCubeExperiment} object which contains summarized rates
for the specified topLevelFeatures.
}
\description{
Title
}
