% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeinAdjustment.R
\name{adjustSynthesisRate}
\alias{adjustSynthesisRate}
\title{Title
TODO}
\usage{
adjustSynthesisRate(spikeinCounts, featureRates, ratio_4sU = 0.1)
}
\arguments{
\item{spikeinCounts}{\code{rCubeExperiment} object with spikeins read counts}

\item{featureRates}{\code{rCubeRates} object with resulting synthesis rates that
should be adjusted}

\item{ratio_4sU}{Percentage of 4sU vs U volume for reverse transcription of spike-ins}
}
\description{
Title
TODO
}
\author{
Carina Demel
}
