% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createJunctionGRangesFromBam.R
\name{createJunctionGRangesFromBam}
\alias{createJunctionGRangesFromBam}
\title{Create de novo junction Annotation from bam files}
\usage{
createJunctionGRangesFromBam(bamFiles, support = 10, ncores = 2,
  BPPARAM = NULL)
}
\arguments{
\item{bamFiles}{A string vector containing all bam files.}

\item{support}{Minimal number of supporting reads.}

\item{ncores}{Number of cores available for parallel computation.}

\item{BPPARAM}{An instance of a \code{BiocParallelParam} class, e.g., 
\code{\link{MulticoreParam}}, \code{\link{SnowParam}}, \code{\link{DoparParam}}.}
}
\value{
Returns a GRanges object with junctions.
}
\description{
Reads coming from sequenced fragments that cannot be 
mapped entirely to a genomic location, can be split, so that they e.g. map 
two exons flanking an intron. The cigar string of these reads usually contains
a 'N' character, where the read cannot be aligned to the genome. This information
can be used to extract possible splice sites.
}
\seealso{
\code{\link{BiocParallelParam}}
}
\author{
Leonhard Wachutka
}
