% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateSizeDispersions.R
\name{estimateSizeDispersions}
\alias{estimateSizeDispersions}
\title{Estimate dispersion values}
\usage{
estimateSizeDispersions(experiment, method = c("DESeqDispMAP", "DESeqDispFit",
  "DESeqDispGeneEst", "Replicate"))
}
\arguments{
\item{experiment}{A \code{rCubeExperiment} object containing feature 
counts for labeled and total RNA-seq samples, see constructor function 
 \code{\link{setupExperiment}}.}

\item{method}{Either "DESeqDispMAP", "DESeqDispFit", "DESeqDispGeneEst", or 
"Replicate" for the type of fitting of dispersions to be used. For the three 
\code{DESeq} methods see also \code{\link[DESeq2]{DESeq}}.}
}
\value{
Returns an updated \code{rCubeExperiment} object with dispersion
estimated included into \code{rowRanges} information
}
\description{
The \code{estimateSizeDispersions} function provides a wrapper function to 
estimate gene-specific dispersion values.
The methods \code{"DESeqDispMAP", "DESeqDispFit", "DESeqDispGeneEst"}
estimate gene-specific dispersion values for total and labeled samples, 
using \code{\link[DESeq2]{DESeq}}.
}
\examples{
data(geneCounts)
geneCounts <- estimateSizeDispersions(geneCounts, method='DESeqDispMAP')
rowRanges(geneCounts)

geneCounts <- estimateSizeDispersions(geneCounts, method='Replicate')
rowRanges(geneCounts)
}
\seealso{
\code{\link[DESeq2]{DESeq}}
}
\author{
Leonhard Wachutka, Carina Demel
}
