% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{estimateSizeFactors}
\alias{estimateSizeFactors}
\alias{calculateNormalizationBySpikeinGLM}
\alias{calculateNormalizationByMean}
\title{Estimation of sample-specific normalization factors}
\usage{
estimateSizeFactors(featureCounts, spikeinCounts, method = c("spikeinGLM",
  "spikeinMean", "jointModel"))

calculateNormalizationBySpikeinGLM(spikeinCounts)

calculateNormalizationByMean(featureCounts, spikeinCounts)
}
\arguments{
\item{featureCounts}{A \code{rCubeExperiment} object containing read
counts for features (genes, junctions, ...), see constructor function 
\code{\link{setupExperiment}}.}

\item{spikeinCounts}{A \code{rCubeExperiment} object containing read 
counts, length and labeling status for spike-ins, as well as sample information,
see constructor function \code{\link{setupExperimentSpikeins}}.}

\item{method}{string specifying which normalization method should be used,
one of ('spikeinGLM','spikeinMean','jointModel').}
}
\value{
Returns an updated \code{rCubeExperiment} object with 
metadata information.

An \code{rCubeExperiment} object for spike-ins with 
updated information in \code{rowData} and \code{colData}.
}
\description{
\code{estimateSizeFactors} provides a wrapper function to estimate 
sample-specific size-factors by three different types.

\code{calculateNormalizationBySpikeinGLM} fits a GLM to spike-in read counts
to extract sequencing depth values and cross-contamination rates for all samples.

\code{calculateNormalizationByMean} ...
}
\examples{
data(geneCounts)
data(spikeinCounts)
### spikeinCounts <- estimateSizeFactors(spikeinCounts, method="spikeinGLM")
geneCounts <- estimateSizeFactors(geneCounts, spikeinCounts, method="spikeinMean")

Leo: do we return feature counts or spike-in counts?
data(spikeinCounts)
spikeinCounts <- calculateNormalizationBySpikeinGLM(spikeinCounts)
}
\author{
Carina Demel
}
