% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{estimateSizeFactors}
\alias{estimateSizeFactors}
\title{Title}
\usage{
estimateSizeFactors(featureCounts, spikeinCounts, method = c("spikeinGLM",
  "spikeinMean", "jointModel"))
}
\arguments{
\item{featureCounts}{rCubeExperiment Object containing read counts for features (genes, junctions, ...)}

\item{spikeinCounts}{rCubeExperiment Object containing read counts for spike-ins}

\item{method}{string specifying which normalization method should be used,
one of ('spikeinGLM','spikeinMean','jointModel')}
}
\value{
rCubeExperiment Object with updated metadata information
}
\description{
Title
}
\examples{

}
\author{
Carina Demel
}
