% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeRates.R
\name{summarizeRates}
\alias{summarizeRates}
\title{Summarize rates for features}
\usage{
summarizeRates(featureRates, topLevelFeatures, by = c("mean", "median"))
}
\arguments{
\item{featureRates}{An \code{rCubeRates} object, see constructor function 
\code{\link{createRResultCubeRates}}.}

\item{topLevelFeatures}{\code{GRanges} annotation}

\item{by}{grouping method, one of 'mean', 'median'}
}
\value{
Returns a \code{rCubeExperiment} object which contains averaged rates
for the specified \code{topLevelFeatures}.
}
\description{
Estimated rates are averaged for features which overlap with a 
top-level feature.
}
\examples{
data(exonRates)
rows <- rowRanges(exonRates)
topLevelFeatures <- reduce(rows)
topLevelFeaturesRates <- summarizeRates(exonRates, topLevelFeatures, by='mean')
}
\seealso{
\code{\link[GenomicRanges]{findOverlaps}}
}
