% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeRates.R
\name{summarizeRates}
\alias{summarizeRates}
\title{Summarize rates for features}
\usage{
summarizeRates(featureRates, topLevelFeatures, by = c("mean", "median"))
}
\arguments{
\item{featureRates}{An \code{rCubeRate object}, see constructor function 
\code{\link{createRResultCubeRates}}.}

\item{topLevelFeatures}{\code{GRanges} annotation}

\item{by}{grouping method ## was this what we intended to do?}
}
\value{
Returns a \code{rCubeExperiment} object which contains summarized rates
for the specified topLevelFeatures.
}
\description{

}
\examples{
data(geneRates)
rows <- rowRanges(geneRates)
topLevelFeatures <- reduce(rows)
topLevelFeaturesRates <- summarizeRates(geneRates, topLevelFeatures, by='mean')
}
